% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auselect_twopp.R
\name{auselect_twopp}
\alias{auselect_twopp}
\title{Retrieve a .csv dataset from the australian_elections GitHub repository.}
\usage{
auselect_twopp(w_ids = T)
}
\arguments{
\item{w_ids}{A boolean value used to determine if the dataset is returned with or without a unique identifier column.
Default is set to TRUE, which returns a dataset with the unique identifier column.}
}
\value{
A filtered dataset with either unique identifier values or no unique identifier values as specified by \code{w_ids}.
}
\description{
\code{auselect_twopp()} downloads and assigns a filtered dataset to a given variable.
}
\details{
The \code{auselect_twopp()} function is used to return a pre-filtered version of either the \code{voting_data_with_ids} or \code{voting_data} datasets.
On call, the function will filter the requested dataset based upon where two party preferred (\code{twoPP}) is true (a value of 1).
The default dataset that is downloaded through this function is the \code{voting_data_with_ids} dataset.
Setting the argument value of the function to \code{w_ids = F} will set the function to instead download the \code{voting_data} dataset.
}
\examples{

\dontrun{
# Return the filtered dataset with unique ids.
twopp_ids <- auselect_twopp()

# Preview dataset
head(twopp_ids)

# Request the filtered dataset without unique ids.
twopp_noids <- auselect_twopp(w_ids = F)

# Preview the dataset.
head(twopp_noids)
}

}
