% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auselect_winners.R
\name{auselect_winners}
\alias{auselect_winners}
\title{Retrieve a .csv dataset from the australian_elections GitHub repository.}
\usage{
auselect_winners(w_ids = T)
}
\arguments{
\item{w_ids}{A boolean value used to determine if the dataset is returned with or without a unique identifier column.
Default is set to TRUE, which returns a dataset with the unique identifier column.}
}
\value{
A filtered dataset with either unique identifier values or no unique identifier values as specified by \code{w_ids}.
}
\description{
`auselect_winners()` downloads and assigns a filtered dataset based upon the election winners to a given variable.
}
\details{
The `auselect_winners()` function is used to return a pre-filtered version of either the `voting_data_with_ids` or `voting_data` datasets.
On call, the function will filter the requested dataset with only election winners (`dummyWinners == 1`).
The default dataset that is downloaded through this function is the `voting_data_with_ids` dataset.
Setting the argument value of the function to `w_ids = F` will set the function to instead download the `voting_data` dataset.
}
\examples{

\dontrun{
# Return the filtered dataset with unique ids.
winners_ids <- auselect_winners()

# Preview dataset
head(winners_ids)

# Request the filtered dataset without unique ids.
winners_noids <- auselect_winners(w_ids = F)

# Preview the dataset.
head(winners_noids)
}

}
