% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{voting_data}
\alias{voting_data}
\title{Elections of all}
\format{A data frame with 1,780 rows and 17 variables that is current as of 24 August 2019:
\describe{
 \item{division}{HAHAHA}
 \item{state}{HAHAHA}
 \item{division_num_enrolled}{HAHAHA}
 \item{division_num_voted}{HAHAHA}
 \item{division_percent_voted}{HAHAHA}
 \item{original_name}{HAHAHA}
 \item{party}{HAHAHA}
 \item{vote_redistribution_round}{HAHAHA}
 \item{votes_count}{HAHAHA}
 \item{votes_share}{HAHAHA}
 \item{swing}{HAHAHA}
 \item{redistributed_vote_count}{HAHAHA}
 \item{redistributed_vote_share}{HAHAHA}
 \item{txt_file}{HAHAHA}
 \item{election_date}{HAHAHA}
 \item{starred}{HAHAHA}
 \item{plus}{HAHAHA}
 \item{capitalised}{HAHAHA}
 \item{twoPP}{HAHAHA}
 \item{number_of_candidates}{HAHAHA}
 \item{winnerDummy}{HAHAHA}
 \item{uniqueID}{HAHAHA}
}}
\usage{
voting_data
}
\description{
This is data about every individual who has been elected to the Australian
House of Representatives or the Senate. Each row is an individual, and the
columns are various pieces of information.
}
\keyword{datasets}
